+++ source/geh_common/src/geh_common/infrastructure/model/current_measurements_column_names.py
@@ -1,0 +1,6 @@
+class CurrentMeasurementsColumnNames:
+    metering_point_id = "metering_point_id"
+    observation_time = "observation_time"
+    quantity = "quantity"
+    quality = "quality"
+    metering_point_type = "metering_point_type"